add: frontend for SingleMovie
